+++ components/MapChart/MapChartHeader/MapChartHeader.jsx
@@ -1,10 +1,10 @@
-import React from 'react'
+import React, { useCallback, useMemo, useState } from 'react'
 import styles from './MapChartHeader.module.scss'
 import MapChartCountries from '../MapChartCountry/MapChartCountry'
 import DropDown from '../../Dropdown/Dropdown'
 import { MapStatistics } from './MapStatistics'
 import '../../../node_modules/mapbox-gl/dist/mapbox-gl.css'
-import { COUNTRY_NAMES } from '../MapConstants'
+import { COUNTRY_CODES, COUNTRY_NAMES } from '../MapConstants'
 
 const MapHeaderControlBar = ({ children }) => {
   return <div className={styles['map-header__control-bar']}>{children}</div>
@@ -20,73 +20,257 @@ const MapLegend = ({
   return (
     <div className={styles['map-header-container__map-legend']}>
       <span>{text}</span>
-      {yearOptions ? (
-        <DropDown
-          dropdownLabel={yearDropdownLabel}
-          options={yearOptions}
-          isDark={true}
-          updateSelection={updateSelection}
-          value={selectedYear}
-        />
-      ) : (
-        <></>
-      )}
+      <DropDown
+        dropdownLabel={yearDropdownLabel}
+        options={yearOptions}
+        isDark={true}
+        updateSelection={updateSelection}
+        value={selectedYear}
+      />
     </div>
   )
 }
 
+const getMapStatistics = (
+  countryData,
+  selectedYear,
+  selectedCountry,
+  selectedProgramType,
+  labels
+) => {
+  let stats = []
+
+  const filteredStats = countryData.filter((country) => {
+    if (selectedCountry === 'All') {
+      return (
+        country.YEAR === selectedYear &&
+        country.LEVEL === 'ALL' &&
+        country.PROGRAM_TYPE === selectedProgramType
+      )
+    }
+    return (
+      country.YEAR === selectedYear &&
+      (country.COUNTRY === selectedCountry || selectedCountry === 'All') &&
+      country.PROGRAM_TYPE === selectedProgramType
+    )
+  })
+
+  if (filteredStats[0]?.PEOPLE_REACHED) {
+    stats.push({
+      title: labels.people,
+      statistic: filteredStats[0].PEOPLE_REACHED
+    })
+  }
+  if (filteredStats[0]?.INVESTMENT) {
+    stats.push({
+      title: labels.invested,
+      statistic: filteredStats[0].INVESTMENT
+    })
+  }
+  if (filteredStats[0]?.PROGRAM_COUNT) {
+    stats.push({
+      title: labels.programs,
+      statistic: filteredStats[0].PROGRAM_COUNT
+    })
+  }
+
+  return stats
+}
+
+//FIXME: after refactor this now shows an option "ALL"
+const getCountryOptions = (countryData, showEmptyPrograms, selectedYear) => {
+  const data = countryData
+    .filter(
+      (country) =>
+        // country.level === 'countries' ||
+        showEmptyPrograms ||
+        (showEmptyPrograms === false && country?.programs?.length > 0) ||
+        selectedYear === country['YEAR']
+    )
+    .reduce(
+      (acc, { COUNTRY: label, programs }) => {
+        if (!label) return acc
+        return [...acc, { label: label, value: label, programs }]
+      },
+      [{ value: 'All', label: 'All' }]
+    )
+  return [...new Map(data.map((i) => [i['label'], i])).values()]
+}
+
+const getProgramOptions = (programData, selectedCountry, selectedYear) => {
+  const countryName = COUNTRY_CODES[selectedCountry]
+  return programData
+    ?.filter(
+      (n) =>
+        n.country === countryName ||
+        selectedCountry === 'All' ||
+        selectedYear === n['YEAR']
+    )
+    .reduce(
+      (acc, { PROGRAM_TYPE }) => {
+        return acc.find((n) => n.value === PROGRAM_TYPE) !== undefined
+          ? acc
+          : [...acc, { label: PROGRAM_TYPE, value: PROGRAM_TYPE }]
+      },
+      [{ value: 'All', label: 'All' }]
+    )
+}
+
+const getYearOptions = (programData) => {
+  const years = programData
+    .map((program) => program['YEAR'])
+    .filter((year) => year !== undefined)
+    .sort((a, b) => b - a)
+  const uniqueYears = [...new Set([...years])]
+  return uniqueYears.map((year) => ({ value: year, label: year }))
+}
+
+//FIXME This data should come from snowflake but is not available a day before go-live
+// there are several programs with more than a single program type, which is currently
+// a different row in the table. We are simply using the first color that comes in
+// for simplicity, but it feels like there is more work to be done here
+const getMarkerColor = (programType) => {
+  const PROGRAM_TYPE_FILL_COLOR = {
+    'Community Development': '#46BB95', //$green
+    'Chronic Emergencies & Fragile Contexts': '#FFBF00',
+    'Crisis Response': '#d00'
+  }
+  return PROGRAM_TYPE_FILL_COLOR[programType]
+}
+
+const getMarkerCoordinates = (
+  programData,
+  selectedYear,
+  selectedCountry,
+  selectedProgramType
+) => {
+  return programData
+    .filter(
+      ({ COUNTRY, PROGRAM_TYPE, YEAR }) =>
+        YEAR === selectedYear &&
+        (COUNTRY === selectedCountry || selectedCountry === 'All') &&
+        (PROGRAM_TYPE === selectedProgramType || selectedProgramType === 'All')
+    )
+    .map((mapData) => {
+      return [
+        mapData['CENTRAL_LONG'],
+        mapData['CENTRAL_LAT'],
+        {
+          ...mapData,
+          fill: getMarkerColor(mapData['PROGRAM_TYPE'])
+        }
+      ]
+    })
+}
+
+const MAP_OPTIONS = {
+  boxZoom: false,
+  doubleClickZoom: false,
+  dragRotate: false,
+  dragPan: false,
+  keyboard: false,
+  scrollZoom: false,
+  touchPitch: false,
+  touchZoomRotate: false,
+  initialViewState: {
+    zoom: 1
+  },
+  minZoom: 1,
+  maxZoom: 8
+}
+
 function MapChartHeader({
   showHeaderControls = true,
+  countryData,
+  programData,
   showMarkers = true,
-  selectedCountry = 'All',
-  selectedProgramType = 'All',
-  onCountryDataLoaded = (evt) => {},
-  selectedYear = 'All',
-  yearOptions = [],
-  onSelectedYearChange = (evt) => {},
-  countryOptions = [],
-  programOptions = [],
-  onSelectedCountryChange = (evt) => {},
-  onSelectedProgramTypeChange = (evt) => {},
-  markerCoordinates = [],
   showEmptyPrograms = false,
   isDark = true,
-  mapStatistics,
-  legendText,
-  countryDropdownLabel,
-  programDropdownLabel,
-  yearDropdownLabel,
-  mapOptions,
+  duration = 1000,
+  padding = 40,
+  labels,
   children,
   ...props
 }) {
-  const mapViewState = {
-    initialViewState: {
-      zoom: 1
-    },
-    minZoom: 1,
-    maxZoom: 8
-  }
+  const [selectedYear, setSelectedYear] = useState()
+  const [selectedCountry, setSelectedCountry] = useState('All')
+  const [selectedProgramType, setSelectedProgramType] = useState('All')
+
+  const countryOptions = useMemo(
+    () => getCountryOptions(countryData, showEmptyPrograms, selectedYear),
+    [countryData, showEmptyPrograms, selectedYear]
+  )
+
+  const programOptions = useMemo(
+    () => getProgramOptions(programData, selectedCountry, selectedYear),
+    [selectedCountry, programData, selectedYear]
+  )
+
+  const yearOptions = useMemo(() => {
+    const options = getYearOptions(programData)
+    setSelectedYear(options[0]?.value || '2021') // always set most current year selected
+
+    return options
+  }, [programData])
+
+  const markerCoordinates = useMemo(() => {
+    return getMarkerCoordinates(
+      programData,
+      selectedYear,
+      selectedCountry,
+      selectedProgramType
+    )
+  }, [selectedYear, selectedCountry, selectedProgramType, programData])
+
+  const mapStatistics = useMemo(() => {
+    return getMapStatistics(
+      // programData,
+      countryData,
+      selectedYear,
+      selectedCountry,
+      selectedProgramType,
+      labels
+    )
+  }, [
+    programData,
+    countryData,
+    selectedYear,
+    selectedCountry,
+    selectedProgramType,
+    labels
+  ])
+
+  const onSelectedYearChange = useCallback((evt) => {
+    setSelectedYear(evt)
+    setSelectedCountry('All')
+  }, [])
+
+  const onSelectedCountryChange = useCallback((evt) => {
+    setSelectedCountry(evt)
+    setSelectedProgramType('All')
+  }, [])
+
+  const onSelectedProgramChange = useCallback((evt) => {
+    setSelectedProgramType(evt)
+  }, [])
 
   return (
     <div className={styles['map-header-container']}>
       <div className={styles['map-container']}>
         <MapChartCountries
           {...props}
-          {...mapViewState}
           color={isDark ? 'dark' : 'light'}
           countryCode={COUNTRY_NAMES[selectedCountry] || 'All'}
-          onCountryDataLoaded={onCountryDataLoaded}
           markerCoordinates={showMarkers ? markerCoordinates : []}
-          mapOptions={mapOptions}
+          mapOptions={MAP_OPTIONS}
         >
           {children}
         </MapChartCountries>
         {showHeaderControls && (
           <MapLegend
-            text={legendText}
+            text={labels.legend}
             selectedYear={selectedYear}
-            yearDropdownLabel={yearDropdownLabel}
+            yearDropdownLabel={labels.year}
             yearOptions={yearOptions}
             updateSelection={onSelectedYearChange}
           />
@@ -98,7 +282,7 @@ function MapChartHeader({
           <div className={styles['map-header__control-bar--first']}>
             <DropDown
               isDark={true}
-              dropdownLabel={countryDropdownLabel}
+              dropdownLabel={labels.country}
               options={countryOptions}
               value={selectedCountry}
               updateSelection={onSelectedCountryChange}
@@ -107,10 +291,10 @@ function MapChartHeader({
           <div className={styles['map-header__control-bar__col--1']}>
             <DropDown
               isDark={true}
-              dropdownLabel={programDropdownLabel}
+              dropdownLabel={labels.program}
               options={programOptions}
-              updateSelection={onSelectedProgramTypeChange}
               value={selectedProgramType}
+              updateSelection={onSelectedProgramChange}
             />
           </div>
           <div className={styles['map-header__control-bar__col--3']}>
